requirements.txt
python-gitlab==2.6.0
PyYAML==5.4.1
